create project - add parallel route
