+++ src/app/@project/default.tsx
@@ -1,3 +1,5 @@
 export default function Default() {
-    return null
+  return null;
 }
+
+
