--- src/app/project-form/page.tsx
@@ -1,20 +1,0 @@
-"use client"
-import ProjectForm, { ProjectI } from "@/components/forms/project.form";
-import { projectAPI } from "@/utils/api";
-import Link from "next/link";
-
-export default function Page() {
-  const onSubmit = async (project: ProjectI) => {
-    console.log('debug', project);
-    try {
-      projectAPI.create({ title: project.title, description: project.description });
-    }
-    catch (error) {
-      console.error(error);
-    }
-  }
-  return <div>
-    <Link href="/">Back</Link>
-    <ProjectForm onSubmit={onSubmit} actionText="Save" />
-  </div>
-}
