--- src/app/project-forms/@project/[...catchAll]/page.tsx
@@ -1,3 +1,0 @@
-export default function CatchAll() {
-    return null
-  }
