--- src/app/project-forms/layout.tsx
@@ -1,14 +1,0 @@
-export default function Layout({
-    children,
-    project,
-  }: {
-    children: React.ReactNode
-    project: React.ReactNode
-  }) {
-    return (
-      <>
-        {children}
-        {project}
-      </>
-    )
-  }
